+++ src/real128.cpp
@@ -8,11 +8,16 @@
 
 #include <mp++/config.hpp>
 
+#include <cassert>
+#include <cstddef>
 #include <ios>
+#include <limits>
+#include <locale>
 #include <ostream>
 #include <sstream>
 #include <stdexcept>
 #include <string>
+#include <vector>
 
 #if defined(MPPP_HAVE_STRING_VIEW)
 
@@ -33,6 +38,7 @@
 #include <quadmath.h>
 
 #include <mp++/detail/utils.hpp>
+#include <mp++/integer.hpp>
 #include <mp++/real128.hpp>
 
 namespace mppp
@@ -172,7 +178,156 @@ real128 nextafter(const real128 &from, const real128 &to)
 // Output stream operator.
 std::ostream &operator<<(std::ostream &os, const real128 &x)
 {
-    detail::float128_stream(os, x.m_value);
+    // Get the stream width.
+    const auto width = os.width();
+
+    // Fetch the stream's flags.
+    const auto flags = os.flags();
+
+    // Check if we are using scientific, fixed or both.
+    const auto scientific = (flags & std::ios_base::scientific) != 0;
+    const auto fixed = (flags & std::ios_base::fixed) != 0;
+    const auto hexfloat = scientific && fixed;
+
+    // Force decimal point character?
+    const auto showpoint = (flags & std::ios_base::showpoint) != 0;
+
+    // Force the plus sign?
+    const bool with_plus = (flags & std::ios_base::showpos) != 0;
+
+    // Uppercase?
+    const bool uppercase = (flags & std::ios_base::uppercase) != 0;
+
+    // Fetch the precision too.
+    auto precision = os.precision();
+    // NOTE: if the precision is negative, reset it
+    // to the default value of 6.
+    if (precision < 0) {
+        precision = 6;
+    }
+
+    // Put together the format string.
+    std::ostringstream oss;
+    oss.exceptions(std::ios_base::failbit | std::ios_base::badbit);
+    oss.imbue(std::locale::classic());
+
+    oss << '%';
+
+    if (showpoint) {
+        oss << '#';
+    }
+
+    if (with_plus) {
+        oss << '+';
+    }
+
+    if (hexfloat) {
+        // NOTE: in hexfloat mode, we want to ignore the precision
+        // setting in the stream object and print the exact representation:
+        // https://en.cppreference.com/w/cpp/locale/num_put/put
+        oss << 'Q';
+    } else {
+        oss << '.' << precision << 'Q';
+    }
+
+    if (hexfloat) {
+        oss << (uppercase ? 'A' : 'a');
+    } else if (scientific) {
+        oss << (uppercase ? 'E' : 'e');
+    } else if (fixed) {
+        // NOTE: in fixed format, the uppercase
+        // setting is ignored for floating-point types:
+        // https://en.cppreference.com/w/cpp/locale/num_put/put
+        oss << 'f';
+    } else {
+        oss << (uppercase ? 'G' : 'g');
+    }
+
+    const auto fmt_str = oss.str();
+
+    // Execute quadmath_snprintf() once to determine the needed buffer size.
+    // NOTE: this returns the number of characters *excluding* the terminator.
+    const auto sz = ::quadmath_snprintf(nullptr, 0, fmt_str.c_str(), x.m_value);
+    if (sz < 0) {
+        // LCOV_EXCL_START
+        throw std::invalid_argument("The quadmath_snprintf() returned an error code");
+        // LCOV_EXCL_STOP
+    }
+    // NOTE: I am not sure if it is possible to get a string of size 0
+    // from quadmath_snprintf(). Let's throw for the time being in such a case,
+    // as zero-length output complicates indexing below.
+    if (sz == 0) {
+        // LCOV_EXCL_START
+        throw std::invalid_argument("The quadmath_snprintf() function returned an empty string");
+        // LCOV_EXCL_STOP
+    }
+    if (sz == std::numeric_limits<int>::max()) {
+        // LCOV_EXCL_START
+        throw std::overflow_error("An overflow condition was detected in the output stream operator of real128");
+        // LCOV_EXCL_STOP
+    }
+
+    // Prepare the char buffer.
+    std::vector<char> buf;
+    // NOTE: need +1 here because the buffer size passed to
+    // quadmath_snprintf() must include the terminator.
+    buf.resize(detail::safe_cast<decltype(buf.size())>(sz + 1));
+
+    // Print out to buf.
+    const auto ret
+        = ::quadmath_snprintf(buf.data(), detail::safe_cast<std::size_t>(buf.size()), fmt_str.c_str(), x.m_value);
+    if (ret < 0) {
+        // LCOV_EXCL_START
+        throw std::invalid_argument("The quadmath_snprintf() returned an error code");
+        // LCOV_EXCL_STOP
+    }
+    // Make sure we wrote the expected number of chars.
+    assert(ret == sz);
+
+    // We are going to do the filling
+    // only if the stream width is larger
+    // than the total size of the string repr.
+    // NOTE: width > ret already implies width >= 0.
+    if (width > ret) {
+        // Determine the fill type.
+        const auto fill = detail::stream_flags_to_fill(flags);
+
+        // Compute how much fill we need.
+        const auto fill_size = detail::safe_cast<decltype(buf.size())>(width - ret);
+
+        // Get the fill character.
+        const auto fill_char = os.fill();
+
+        switch (fill) {
+            case 1:
+                // Left fill: fill characters at the end.
+                // NOTE: minus 1 because of the terminator.
+                buf.insert(buf.end() - 1, fill_size, fill_char);
+                break;
+            case 2:
+                // Right fill: fill characters at the beginning.
+                buf.insert(buf.begin(), fill_size, fill_char);
+                break;
+            default: {
+                assert(fill == 3);
+
+                // Internal fill: the fill characters are always after the sign (if present).
+                const auto delta = static_cast<int>(buf[0] == '+' || buf[0] == '-');
+                buf.insert(buf.begin() + delta, fill_size, fill_char);
+                break;
+            }
+        }
+    }
+
+    // Write from buf into the stream.
+    // NOTE: minus 1 because of the terminator.
+    os.write(buf.data(), detail::safe_cast<std::streamsize>(buf.size() - 1u));
+
+    // Reset the stream width to zero, like the operator<<() does for builtin types.
+    // https://en.cppreference.com/w/cpp/io/manip/setw
+    // Do it here so we ensure we don't alter the state of the stream until the very end.
+    os.width(0);
+
     return os;
 }